Entrada de primeros metodos y arreglo de tablas
Arregladas tablas
-Se cambiaron valores int por long para coincidir con base de datos
-Se modificó el dbcontext para las constraints de algunas tablas
-Se agregaron metodos httprequest para la creación de usuario y post, traer usuario por id

+++ SMedia/SMedia/Models/Message.cs
@@ -7,10 +7,10 @@ namespace SMedia.Models
 {
     public class Message
     {
-        public int Id { get; set; }
+        public long Id { get; set; }
         public string Content { get; set; }
-        public int SenderId { get; set; }
-        public int ReceiverId { get; set; }
+        public long SenderId { get; set; }
+        public long ReceiverId { get; set; }
         public DateTime Date { get; set; }
 
         public virtual User Sender { get; set; }
